Move custom_component.py to a new pipeline dir
