+++ rest_api/controller/__init__.py
@@ -1,1 +1,1 @@
-from rest_api import custom_component  # this import is required for the Custom Components to be registered
+from rest_api.pipeline import custom_component  # this import is required for the Custom Components to be registered
